Move Exception defs to an exceptions package
This package colocates all exceptions that can be thrown by SALVE. Not only
does this serve to make the code more self-documenting, as the thrown
exceptions are now easily findable, but it also ensures that none of them are
being defined in package __init__s
