+++ salve/exceptions/__init__.py
@@ -1,0 +1,6 @@
+from salve.exceptions.base import SALVEException
+
+from salve.exceptions.parsing import ParsingException, TokenizationException
+
+from salve.exceptions.block import BlockException
+from salve.exceptions.action import ActionException
