+++ salve/exceptions/base.py
